activeCalls/delete-active-call/requirements.txt
fdk>=0.1.50
oci
python-dotenv